TableRender takes Table as attribute
